(#112) Wait for media element to be ready before generating waveform
This commit adds code to wait for a 'canplay' event from the media
element, so that the currentSrc attribute is valid.

This fixes a race condition where an <audio> element that contains
one or more <source> elements is added to a page, and Peaks.init()
is called before the <audio> element has selected which <source>
to use.