Add drag start and drag end events to segments
When an editable segment has its in or out marker dragged, event
listeners have no way of knowing which marker is being dragged.
This change adds a drag start and drag end event which pass the
segment and a boolean inMarker that allows users to track which
marker is being dragged.