Revised mousewheel handling
Added captureVerticalScroll option to view.setWheelMode().

This option makes the zoomview respond to all mousewheel / touchpad
events. It means that the user no longer has to press Shift to
scroll the waveform, but it also means that using the mousewheel
while the pointer is over the waveform will no longer scroll the
page.

See #454