Disable Worker when using an AudioBuffer
The external player demo page (that uses Tone.js) was not playing audio.

This was caused by the waveform-data library, which calls getChannelData()
on the AudioBuffer and transfers the channels' ArrayBuffers to the Worker.

It's unclear to me why this prevents audio from being played, as I'd assume
the AudioBuffer should be unaffected by this, but setting the disable_worker
option when calling WaveformData.createFromAudio() avoids the problem.

Resolves #451