Improved segment dragging
Move the segment being dragged to the top of the z-order, so that it
remains on top of any other segments that the handle is dragged over.

This isn't perfect, as the other segment's handle is still covered
by the segment that was dragged, so I made the segment click handler
bring the segment to the top.